+++ lib/src/abacatepay.dart
@@ -1,5 +1,6 @@
 import 'package:abacatepay/src/client/abacatepay_client.dart';
 import 'package:abacatepay/src/datasources/billing/billing_datasource.dart';
+import 'package:abacatepay/src/datasources/coupon/coupon_datasource.dart';
 import 'package:abacatepay/src/datasources/customers/customers_datasource.dart';
 import 'package:abacatepay/src/datasources/pixQrCode/pix_qr_code_datasource.dart';
 
@@ -14,6 +15,9 @@ final class AbacatePay {
   /// Métodos relacionados ao QrCode PIX.
   late final AbacatePayPixQrCode pixQrCode;
 
+  /// Métodos relacionados a cupons.
+  late final AbacatePayCoupon coupon;
+
   late final String _apiKey;
   late final int _apiVersion;
 
@@ -26,5 +30,6 @@ final class AbacatePay {
     billing = AbacatePayBilling(client);
     customers = AbacatePayCustomer(client);
     pixQrCode = AbacatePayPixQrCode(client);
+    coupon = AbacatePayCoupon(client);
   }
 }