+++ lib/src/models/enums/discount_kind_enum.dart
@@ -1,0 +1,10 @@
+import 'package:abacatepay/src/models/constants/abacatepay_constants.dart';
+
+enum DiscountKind {
+  percentage(AbacatePayConstants.percentageDiscountEnum),
+  fixed(AbacatePayConstants.fixedDiscountEnum);
+
+  final String apiKey;
+
+  const DiscountKind(this.apiKey);
+}